Rewrite roles implementation
This piggybacks on top of Rake::RemoteTask's role definitions,
but overrides the behavior. This requires some monkey-patching.
An improvement for a future version of Fezzik is to subclass
RemoteTask (or rewrite its functionality) and use it throughout.

This implementation also obviates the need to overwrite remote_task,
which is excellent.